backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
feedparser==6.0.10
httpx==0.25.2
python-dotenv==1.0.0
vaderSentiment==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
apscheduler==3.10.4
tenacity==8.2.3
orjson==3.9.10
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
ruff==0.1.8
openai>=1.43.0
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.66

